pyproject.toml
[project]
name = "srt-reservation"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Seonghyeon Cho", email = "seonghyeoncho96@gmail.com" }
]
dependencies = [
    "selenium>=4.23.1",
    "webdriver-manager>=4.0.2",
    "python-telegram-bot>=21.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/srt_reservation"]
